+++ controller/settingscontroller.php
@@ -128,7 +128,6 @@ class SettingsController extends Controller {
 			"feedback" => $this->config->getCustomizationFeedback(),
 			"forcesave" => $this->config->getCustomizationForcesave(),
 			"help" => $this->config->getCustomizationHelp(),
-			"toolbarNoTabs" => $this->config->getCustomizationToolbarNoTabs(),
 			"successful" => $this->config->settingsAreSuccessful(),
 			"plugins" => $this->config->getCustomizationPlugins(),
 			"macros" => $this->config->getCustomizationMacros(),
@@ -218,7 +217,6 @@ class SettingsController extends Controller {
 	 * @param bool $feedback - display feedback
 	 * @param bool $forcesave - forcesave
 	 * @param bool $help - display help
-	 * @param bool $toolbarNoTabs - display toolbar tab
 	 * @param string $reviewDisplay - review viewing mode
 	 * @param string $theme - default theme mode
 	 * @param string $unknownAuthor - display unknown author
@@ -239,7 +237,6 @@ class SettingsController extends Controller {
 		$feedback,
 		$forcesave,
 		$help,
-		$toolbarNoTabs,
 		$reviewDisplay,
 		$theme,
 		$unknownAuthor
@@ -257,7 +254,6 @@ class SettingsController extends Controller {
 		$this->config->setCustomizationFeedback($feedback);
 		$this->config->setCustomizationForcesave($forcesave);
 		$this->config->setCustomizationHelp($help);
-		$this->config->setCustomizationToolbarNoTabs($toolbarNoTabs);
 		$this->config->setCustomizationReviewDisplay($reviewDisplay);
 		$this->config->setCustomizationTheme($theme);
 		$this->config->setUnknownAuthor($unknownAuthor);