+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - edit xlsb
 - customization_goback setting removed
 - there are only system, light and dark themes
+- toolbarNoTabs settings removed
 
 ## 9.9.1
 ## Added